Tweaked Server routes
