+++ package.json
@@ -51,9 +51,7 @@
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.5.1",
     "nodemon": "^2.0.2",
-
     "protractor": "^5.4.3",
-
     "prettier": "^1.19.1",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
